third_party/blink/web_tests/flag-specific/disable-layout-ng/fast/forms/file/display-none-upload-button-expected.txt

FAIL document.querySelector('input').offsetHeight > 10 should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE

Tests that a display:none upload button works. 
